kairos-contracts: factor out contract-utils
