+++ kairos-contracts/demo-contract/contract-utils/src/constants.rs
