+++ kairos-contracts/demo-contract/contract-utils/src/lib.rs
@@ -1,6 +1,10 @@
+#![no_std]
+pub mod constants;
+
+use casper_event_standard::casper_types::Key;
 use casper_event_standard::Event;
+
 extern crate alloc;
-use casper_types::Key;
 
 #[derive(Event)]
 pub struct Deposit {
